+++ src/platform/packages/private/kbn-visualization-types-and-defaults/index.ts
@@ -1,0 +1,49 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+/**
+ * Export all types
+ * To simplify the internal organizazion types are grouped in two categories:
+ * - `expression` types: there are the types used in the expression language and often by visualize directly
+ * - `types` types: there are the types used in the lens code (both editor and embeddable)
+ */
+
+export * from './src/types';
+export * from './src/datasources/types';
+export * from './src/visualizations/types';
+export * from './src/visualizations/datatable/expression_types';
+export * from './src/visualizations/datatable/types';
+export * from './src/visualizations/gauge/expression_types';
+export * from './src/visualizations/gauge/types';
+export * from './src/visualizations/heatmap/expression_types';
+export * from './src/visualizations/heatmap/types';
+export * from './src/visualizations/legacy_metric/expression_types';
+export * from './src/visualizations/legacy_metric/types';
+export * from './src/visualizations/metric/expression_types';
+export * from './src/visualizations/metric/types';
+export * from './src/visualizations/partition/expression_types';
+export * from './src/visualizations/partition/types';
+export * from './src/visualizations/tagcloud/expression_types';
+export * from './src/visualizations/tagcloud/types';
+export * from './src/visualizations/xy/expression_types';
+export * from './src/visualizations/xy/types';
+
+/**
+ * Constants are used by both the expression language and the lens code
+ */
+export * from './src/constants';
+export * from './src/datasources/constants';
+export * from './src/visualizations/constants';
+export * from './src/visualizations/gauge/constants';
+export * from './src/visualizations/heatmap/constants';
+export * from './src/visualizations/legacy_metric/constants';
+export * from './src/visualizations/metric/constants';
+export * from './src/visualizations/partition/constants';
+export * from './src/visualizations/tagcloud/constants';
+export * from './src/visualizations/xy/constants';
