+++ src/platform/packages/private/kbn-visualization-types-and-defaults/kibana.jsonc
@@ -1,0 +1,7 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/visualization-types-and-defaults",
+  "owner": "@elastic/kibana-visualizations",
+  "group": "platform",
+  "visibility": "private"
+}
