+++ src/platform/packages/private/kbn-visualization-types-and-defaults/src/constants.ts
@@ -1,0 +1,19 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+export const EMPTY_SIZE_RADIUS = {
+  SMALL: 0.3,
+  MEDIUM: 0.54,
+  LARGE: 0.7,
+} as const;
+
+export const ROW_HEIGHT_MODE = {
+  auto: 'auto',
+  custom: 'custom',
+} as const;
