+++ src/platform/packages/private/kbn-visualization-types-and-defaults/src/visualizations/heatmap/constants.ts
@@ -7,6 +7,8 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
+import { IconChartHeatmap } from '@kbn/chart-icons';
+import { i18n } from '@kbn/i18n';
 import {
   RequiredPaletteParamTypes,
   FIXED_PROGRESSION,
@@ -16,7 +18,34 @@ import {
   DEFAULT_COLOR_STEPS,
 } from '@kbn/coloring';
 
-export const defaultPaletteParams: RequiredPaletteParamTypes = {
+export const LENS_HEATMAP_ID = 'lnsHeatmap';
+
+export const LENS_HEATMAP_CHART_SHAPES = {
+  HEATMAP: 'heatmap',
+} as const;
+
+export const LENS_HEATMAP_CHART_NAMES = {
+  heatmap: {
+    shapeType: LENS_HEATMAP_CHART_SHAPES.HEATMAP,
+    icon: IconChartHeatmap,
+    label: i18n.translate('xpack.lens.heatmap.heatmapLabel', {
+      defaultMessage: 'Heat map',
+    }),
+  },
+};
+
+export const LENS_HEATMAP_GROUP_ID = {
+  X: 'x',
+  Y: 'y',
+  CELL: 'cell',
+} as const;
+
+export const EXPRESSION_HEATMAP_NAME = 'heatmap';
+export const EXPRESSION_HEATMAP_LEGEND_NAME = 'heatmap_legend';
+export const EXPRESSION_HEATMAP_GRID_NAME = 'heatmap_grid';
+
+export const LENS_HEATMAP_DEFAULT_PALETTE_NAME = 'temperature';
+export const LENS_HEATMAP_DEFAULT_PALETTE_PARAMS: RequiredPaletteParamTypes = {
   name: 'temperature',
   reverse: false,
   rangeType: 'percent',
