+++ src/platform/packages/private/kbn-visualization-types-and-defaults/src/visualizations/heatmap/expression_types.ts
@@ -1,0 +1,81 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import { Position } from '@elastic/charts';
+import { ExpressionValueVisDimension, LegendSize } from '@kbn/visualizations-plugin/common';
+import { CustomPaletteState } from '@kbn/charts-plugin/common';
+import { PaletteOutput } from '@kbn/coloring';
+import { LayerType } from '../types';
+import { EXPRESSION_HEATMAP_GRID_NAME, EXPRESSION_HEATMAP_LEGEND_NAME } from './constants';
+
+export type HeatmapChartShapes = 'heatmap';
+
+export interface HeatmapLegendConfigResult {
+  /**
+   * Flag whether the legend should be shown. If there is just a single series, it will be hidden
+   */
+  isVisible: boolean;
+  /**
+   * Position of the legend relative to the chart
+   */
+  position: Position;
+  /**
+   * Defines the number of lines per legend item
+   */
+  maxLines?: number;
+  /**
+   * Defines if the legend items should be truncated
+   */
+  shouldTruncate?: boolean;
+  /**
+   * Exact legend width (vertical) or height (horizontal)
+   * Limited to max of 70% of the chart container dimension Vertical legends limited to min of 30% of computed width
+   */
+  legendSize?: LegendSize;
+  // type flag
+  type: typeof EXPRESSION_HEATMAP_LEGEND_NAME;
+}
+
+export interface HeatmapGridConfigResult {
+  // grid
+  strokeWidth?: number;
+  strokeColor?: string;
+  // cells
+  isCellLabelVisible: boolean;
+  // Y-axis
+  isYAxisLabelVisible: boolean;
+  isYAxisTitleVisible: boolean;
+  yTitle?: string;
+  // X-axis
+  isXAxisLabelVisible: boolean;
+  xAxisLabelRotation?: number;
+  isXAxisTitleVisible: boolean;
+  xTitle?: string;
+  // type flag
+  type: typeof EXPRESSION_HEATMAP_GRID_NAME;
+}
+
+export interface HeatmapExpressionLayerState {
+  layerId: string;
+  layerType: LayerType;
+  xAccessor?: string | ExpressionValueVisDimension;
+  yAccessor?: string | ExpressionValueVisDimension;
+  valueAccessor?: string | ExpressionValueVisDimension;
+  splitRowAccessor?: string | ExpressionValueVisDimension;
+  splitColumnAccessor?: string | ExpressionValueVisDimension;
+  palette?: PaletteOutput<CustomPaletteState>;
+  shape: HeatmapChartShapes;
+  percentageMode?: boolean;
+  lastRangeIsRightOpen?: boolean;
+  showTooltip?: boolean;
+  highlightInHover?: boolean;
+  legend: HeatmapLegendConfigResult;
+  gridConfig: HeatmapGridConfigResult;
+  ariaLabel?: string;
+}
