+++ src/platform/packages/private/kbn-visualization-types-and-defaults/src/visualizations/legacy_metric/constants.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+export const LEGACY_METRIC_LABEL_POSITION = {
+  BOTTOM: 'bottom',
+  TOP: 'top',
+} as const;
