+++ src/platform/packages/private/kbn-visualization-types-and-defaults/src/visualizations/legacy_metric/types.ts
@@ -1,0 +1,27 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import type { Simplify } from '../../utils';
+import type { LayerType } from '../types';
+import type {
+  LegacyMetricAlignment,
+  LegacyMetricArguments,
+  LegacyMetricLabelPositionType,
+} from './expression_types';
+
+export type LegacyMetricState = Simplify<
+  Partial<Pick<LegacyMetricArguments, 'autoScaleMetricAlignment' | 'colorMode' | 'palette'>> & {
+    layerId: string;
+    accessor?: string;
+    layerType: LayerType;
+    titlePosition?: LegacyMetricLabelPositionType;
+    size?: string;
+    textAlign?: LegacyMetricAlignment;
+  }
+>;
