+++ src/platform/packages/private/kbn-visualization-types-and-defaults/src/visualizations/partition/constants.ts
@@ -1,0 +1,55 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import { $Values } from '@kbn/utility-types';
+
+export const PieChartTypes = {
+  PIE: 'pie',
+  DONUT: 'donut',
+  TREEMAP: 'treemap',
+  MOSAIC: 'mosaic',
+  WAFFLE: 'waffle',
+} as const;
+
+export type PieChartType = $Values<typeof PieChartTypes>;
+
+export interface SharedPieLayerState {
+  metrics: string[];
+  primaryGroups: string[];
+  secondaryGroups?: string[];
+  allowMultipleMetrics?: boolean;
+  colorsByDimension?: Record<string, string>;
+  collapseFns?: Record<string, CollapseFunction>;
+  numberDisplay: NumberDisplayType;
+  categoryDisplay: CategoryDisplayType;
+  legendDisplay: LegendDisplayType;
+  legendPosition?: Position;
+  legendStats?: PartitionLegendValue[];
+  nestedLegend?: boolean;
+  percentDecimals?: number;
+  emptySizeRatio?: number;
+  legendMaxLines?: number;
+  legendSize?: LegendSize;
+  truncateLegend?: boolean;
+  colorMapping?: ColorMapping.Config;
+}
+
+export type PieLayerState = SharedPieLayerState & {
+  layerId: string;
+  layerType: LayerType;
+};
+
+export interface PieVisualizationState {
+  shape: $Values<typeof PieChartTypes>;
+  layers: PieLayerState[];
+  /**
+   * @deprecated use `layer.colorMapping` config
+   */
+  palette?: PaletteOutput;
+}
