+++ src/platform/packages/private/kbn-visualization-types-and-defaults/src/visualizations/types.ts
@@ -1,0 +1,17 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import { $Values } from '@kbn/utility-types';
+import { CategoryDisplay, LegendDisplay, NumberDisplay, layerTypes } from './constants';
+
+export type CategoryDisplayType = $Values<typeof CategoryDisplay>;
+export type NumberDisplayType = $Values<typeof NumberDisplay>;
+export type LegendDisplayType = $Values<typeof LegendDisplay>;
+
+export type LayerType = (typeof layerTypes)[keyof typeof layerTypes];
