+++ src/platform/plugins/shared/chart_expressions/expression_gauge/common/types/expression_functions.ts
@@ -6,63 +6,21 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-
-import { $Values } from '@kbn/utility-types';
-import type { PaletteOutput, CustomPaletteParams } from '@kbn/coloring';
-import {
+import type { GaugeColorMode, GaugeExpressionState } from '@kbn/visualization-types-and-defaults';
+import type { PaletteOutput } from '@kbn/coloring';
+import type {
   Datatable,
   DefaultInspectorAdapters,
   ExecutionContext,
   ExpressionFunctionDefinition,
   ExpressionValueRender,
 } from '@kbn/expressions-plugin/common';
-import { ExpressionValueVisDimension } from '@kbn/visualizations-plugin/common';
-import { CustomPaletteState } from '@kbn/charts-plugin/common';
+import type { CustomPaletteState } from '@kbn/charts-plugin/common';
 import type { MakeOverridesSerializable, Simplify } from '@kbn/chart-expressions-common/types';
 import type { GoalProps } from '@elastic/charts';
-import {
-  EXPRESSION_GAUGE_NAME,
-  GAUGE_FUNCTION_RENDERER_NAME,
-  GaugeShapes,
-  GaugeTicksPositions,
-  GaugeLabelMajorModes,
-  GaugeColorModes,
-  GaugeCentralMajorModes,
-} from '../constants';
-export type GaugeColorMode = $Values<typeof GaugeColorModes>;
-export type GaugeShape = $Values<typeof GaugeShapes>;
-export type GaugeLabelMajorMode = $Values<typeof GaugeLabelMajorModes>;
-export type GaugeCentralMajorMode = $Values<typeof GaugeCentralMajorModes>;
-export type GaugeTicksPosition = $Values<typeof GaugeTicksPositions>;
+import { EXPRESSION_GAUGE_NAME, GAUGE_FUNCTION_RENDERER_NAME } from '../constants';
 
-export interface GaugeState {
-  metric?: string | ExpressionValueVisDimension;
-  min?: string | ExpressionValueVisDimension;
-  max?: string | ExpressionValueVisDimension;
-  goal?: string | ExpressionValueVisDimension;
-  ticksPosition: GaugeTicksPosition;
-  labelMajorMode: GaugeLabelMajorMode;
-  labelMajor?: string;
-  labelMinor?: string;
-  colorMode?: GaugeColorMode;
-  palette?: PaletteOutput<CustomPaletteParams>;
-  shape: GaugeShape;
-  respectRanges?: boolean;
-  commonLabel?: string;
-  /**
-   * @deprecated Use `labelMajorMode` instead
-   */
-  centralMajorMode?: GaugeCentralMajorMode;
-  /**
-   * @deprecated Use `labelMajor` instead
-   */
-  centralMajor?: string;
-  /**
-   * This field is deprecated and will be removed in a future release
-   * @deprecated
-   */
-  percentageMode?: boolean;
-}
+export type GaugeState = GaugeExpressionState;
 
 export type GaugeArguments = GaugeState & {
   colorMode: GaugeColorMode;
