+++ src/platform/plugins/shared/chart_expressions/expression_legacy_metric/common/types/expression_functions.ts
@@ -17,19 +17,23 @@ import {
   Style,
 } from '@kbn/expressions-plugin/common';
 import { ExpressionValueVisDimension } from '@kbn/visualizations-plugin/common';
-import { ColorMode, CustomPaletteState } from '@kbn/charts-plugin/common';
-import { VisParams, visType, LabelPositionType, MetricAlignment } from './expression_renderers';
+import type { ColorMode, CustomPaletteState } from '@kbn/charts-plugin/common';
+import type {
+  LegacyMetricAlignment,
+  LegacyMetricLabelPositionType,
+} from '@kbn/visualization-types-and-defaults';
+import { VisParams, visType } from './expression_renderers';
 import { EXPRESSION_METRIC_NAME } from '../constants';
 
 export interface MetricArguments {
-  autoScaleMetricAlignment?: MetricAlignment;
+  autoScaleMetricAlignment?: LegacyMetricAlignment;
   percentageMode: boolean;
   colorMode: ColorMode;
   showLabels: boolean;
   palette?: PaletteOutput<CustomPaletteState>;
   font: Style;
   labelFont: Style;
-  labelPosition: LabelPositionType;
+  labelPosition: LegacyMetricLabelPositionType;
   metric: Array<ExpressionValueVisDimension | string>;
   bucket?: ExpressionValueVisDimension | string;
   colorFullBackground: boolean;
