+++ src/platform/plugins/shared/visualizations/common/index.ts
@@ -33,6 +33,7 @@ export {
   getAccessor,
   getFormatByAccessor,
   getStopsWithColorsFromRanges,
+  type InspectorLogTable,
 } from './utils';
 export type {
   Arguments,
