+++ tsconfig.base.json
@@ -2166,6 +2166,8 @@
       "@kbn/vis-type-vislib-plugin/*": ["src/platform/plugins/private/vis_types/vislib/*"],
       "@kbn/vis-type-xy-plugin": ["src/platform/plugins/private/vis_types/xy"],
       "@kbn/vis-type-xy-plugin/*": ["src/platform/plugins/private/vis_types/xy/*"],
+      "@kbn/visualization-types-and-defaults": ["src/platform/packages/private/kbn-visualization-types-and-defaults"],
+      "@kbn/visualization-types-and-defaults/*": ["src/platform/packages/private/kbn-visualization-types-and-defaults/*"],
       "@kbn/visualization-ui-components": ["src/platform/packages/shared/kbn-visualization-ui-components"],
       "@kbn/visualization-ui-components/*": ["src/platform/packages/shared/kbn-visualization-ui-components/*"],
       "@kbn/visualization-utils": ["src/platform/packages/shared/kbn-visualization-utils"],
